fixxed big transformer bug
transformer length prediction 2 layer linear neural net was taking as input size "latent_mem" but was outputing size "latent mem". This resulted in errors when the input shape was different from the latent shape.
Solution was to change the output shape of this network to match the input shape, "d_model" here.